+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog] and this project adheres to [Semantic Versioning].
 
-## [0.2.2] - 2021-01-25
+## [1.0.0] - 2021-01-25
+
+> From this version on Mautic 2 will not work anymore
+
+### Fixed
+
+- Plugin now functional for Mautic 3.3
+
+## [0.2.3] - 2021-01-25
 
 ### Fixed
 
@@ -48,6 +56,7 @@ The format is based on [Keep a Changelog] and this project adheres to [Semantic 
 
 Inial release 🎉
 
+[1.0.0]: https://github.com/bastolen/mautic-message-extension/releases/tag/1.0.0
 [0.2.3]: https://github.com/bastolen/mautic-message-extension/releases/tag/0.2.3
 [0.2.2]: https://github.com/bastolen/mautic-message-extension/releases/tag/0.2.2
 [0.2.1]: https://github.com/bastolen/mautic-message-extension/releases/tag/0.2.1